+++ cocos/3d/framework/mesh-renderer.ts
@@ -307,6 +307,9 @@ export class MeshRenderer extends ModelRenderer {
     @serializable
     protected _reflectionProbeBlendWeight = 0;
 
+    @serializable
+    protected _enabledGlobalStandardSkinObject = false;
+
     protected _reflectionProbeDataMap: Texture2D | null = null;
 
     // @serializable
@@ -480,6 +483,29 @@ export class MeshRenderer extends ModelRenderer {
 
     set enableMorph (value) {
         this._enableMorph = value;
+    }
+
+    /**
+     * @en local shadow normal bias for real time lighting.
+     * @zh 实时光照下模型局部的阴影法线偏移。
+     */
+    @type(CCBoolean)
+    @tooltip('i18n:model.standard_skin_model')
+    @disallowAnimation
+    get isGlobalStandardSkinObject () {
+        return this._enabledGlobalStandardSkinObject;
+    }
+
+    set isGlobalStandardSkinObject (val) {
+        cclegacy.director.root.pipeline.pipelineSceneData.standardSkinModel = val ? this : null;
+        this._enabledGlobalStandardSkinObject = val;
+    }
+
+    /**
+     * @engineInternal
+     */
+    public clearGlobalStandardSkinObjectFlag () {
+        this._enabledGlobalStandardSkinObject = false;
     }
 
     protected _modelType: typeof scene.Model;
@@ -1225,6 +1251,12 @@ export class MeshRenderer extends ModelRenderer {
 
     private _uploadSubMeshShapesWeights (subMeshIndex: number) {
         this._morphInstance?.setWeights(subMeshIndex, this._subMeshShapesWeights[subMeshIndex]);
+    }
+
+    private _updateStandardSkin () {
+        if (this._enabledGlobalStandardSkinObject) {
+            cclegacy.director.root.pipeline.pipelineSceneData.standardSkinModel = this;
+        }
     }
 }
